fix: user creation and updating
Previously, the app would crash when attempting to create a new user. This was because as soon as the create function ran the `createUserWithEmailAndPassword` function, the watchAuthState function, which listened to changes in auth, would assume the app has been fully logged in before user data was populated within firestore. This made the app switch to the Loading and then the Home screen before the data was created. The data would be fetched, and being missing, would cause the app to crash before the data even existed.

To resolve this, a new "isUpdatingUser" store is created. This store is set to true when a user is being updated, and false when finished. When a user is created, the loading screen now checks that store before switching to the home screen.

This code also fixes updating a user when the data doesn't exist in the database. Previously, the code would use the user's id from userSlice to update data, but in the case that the user isn't in the db, this wouldn't exist. To resolve this, it picks between that id and firebase-auth's user id depending on which one is available.

This one took a while, but it workssss.